requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.1
sqlalchemy==2.1.3
python-multipart==0.0.6
pydantic==2.5.1
PyMuPDF==1.22.5
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai>=1.40.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
requests
python-dotenv
